research-assistant/requirements.txt
streamlit
openai
sentence-transformers
chromadb
python-dotenv
pypdf
langchain